backend/requirements.txt
fastapi
uvicorn[standard]
motor
dnspython
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
